Time: 244 ms (42.84%), Space: 17 MB (21.45%) - LeetHub

+++ 1463-cherry-pickup-ii/1463-cherry-pickup-ii.cpp
@@ -36,10 +36,10 @@ public:
         for(int i=n-2; i>=0; i--){
             for(int j1=0;j1<m; j1++){
                 for(int j2=0; j2<m; j2++){
-                   long long  int maxi=-1e8;
+                   int maxi=-1e8;
                     for(int di=-1; di<=1;di++){
                     for(int dj=-1; dj<=1;dj++){
-                  long long  int ans;
+                   int ans;
                     if(j1==j2) ans= grid[i][j1] ;
                     else ans=grid[i][j1]+grid[i][j2];
                         if(j1+di>=0&& j1+di <m&& j2+dj>=0&&j2+dj<m){
